Refactor: Move utility classes to core package
This commit restructures the project by moving utility classes and functions, such as `Error`, `Result`, and `CoinCodeToResource`, to a new core package.

Specifically, these classes are relocated to the `core.domain.util` and `core.presentation.util` sub-packages within the core package.

Additionally, a new `NetworkError` enum class is introduced in the `core.domain.util` package to represent different types of network errors.

This refactoring improves the organization of the project and promotes code reuse.